+++ man-roxygen/DbmsDetails.R
@@ -9,6 +9,7 @@
 #'                             \item {"bigquery" for Google BigQuery}
 #'                             \item {"sqlite" for SQLite}
 #'                             \item {"sqlite extended" for SQLite with extended types (DATE and DATETIME)}
+#'                             \item {"spark" for Spark}
 #'                           }
 #'
 #' @param user               The user name used to access the server.
@@ -113,7 +114,13 @@
 #' }
 #' SQLite:
 #' \itemize{
-#'   \item \code{server}. The path to the SQLIte file
+#'   \item \code{server}. The path to the SQLIte file.
+#' }
+#' Spark:
+#' \itemize{
+#'   \item \code{connectionString}. The connection string (e.g. starting with 'jdbc:spark://my-org.dev.cloud.databricks.com...').
+#'   \item \code{user}. The user name used to access the server.
+#'   \item \code{password}. The password for that user.
 #' }
 #'
 #' To be able to use Windows authentication for SQL Server (and PDW), you have to install the JDBC